Cambios 13/04/2019
-Implementado gráfico con historial de cálculos en Perfil

+++ src/CalculadoraRCV/app/src/main/java/ipo2/es/calculadorarcv/dominio/CalculoRCV.java
@@ -93,7 +93,7 @@ public class CalculoRCV implements Serializable, ConstantesFactores {
         this.idUser = id;
     }
 
-    public double getScore() {
+    public int getScore() {
         return score;
     }
